restructures file system in preparation for Redux
