+++ package.json
@@ -51,7 +51,9 @@
     "react-bem-helper": "^1.2.0",
     "react-dom": "^15.4.2",
     "react-masonry-component": "^5.0.3",
-    "react-router": "^3.0.2"
+    "react-redux": "^5.0.2",
+    "react-router": "^3.0.2",
+    "redux": "^3.6.0"
   },
   "scripts": {
     "start": "node scripts/start.js",
