+++ src/components/Landing/Introduction/Projects/Projects.scss
@@ -1,0 +1,29 @@
+.c-projects {
+  margin: 0 auto;
+  width: 100%;
+
+  &__title {
+    @extend .c--title;
+
+    margin: 10px;
+    padding: 0;
+    letter-spacing: 0.15em;
+    font-size: 1.5em;
+
+  }
+
+  &__masonry:after {
+    content: '';
+    display: block;
+    clear: both;
+  }
+}
+
+@media screen and (max-width: $one-columns) {
+  .c-projects {
+    &__masonry {
+      margin: 0 auto;
+      width: 100%;
+    }
+  }
+}
